+++ app/actions/infoPanelActions.js
@@ -1,0 +1,9 @@
+import * as types from './actionTypes';
+
+const modifyInfoPanelDisplay = () => (dispatch) => {
+  dispatch({
+    type: types.MODIFY_INFO_PANEL_DISPLAY,
+  });
+};
+
+export default modifyInfoPanelDisplay;