+++ app/tests/components/Header.test.jsx
@@ -7,9 +7,13 @@ import Header from '../../components/App/Header';
 Enzyme.configure({ adapter: new Adapter() });
 const MockFn = jest.fn().mockImplementation(() => {});
 
-const getComponentWithDefaultProps = (onReloadTopicsRequest = () => { MockFn(); }) => (
+const getComponentWithDefaultProps = (
+  onReloadTopicsRequest = () => { MockFn(); },
+  onModifyInfoPanelDisplayRequest = () => { MockFn(); }
+  ) => (
   <Header
     onReloadTopicsRequest={onReloadTopicsRequest}
+    onModifyInfoPanelDisplayRequest={onModifyInfoPanelDisplayRequest}
   />
     );
 
@@ -17,12 +21,12 @@ describe('<Header />', () => {
     it('renders default state', () => {
         const wrapper = shallow(getComponentWithDefaultProps());
         expect(wrapper.find('span')).toHaveLength(1);
-        expect(wrapper.find('button')).toHaveLength(2);
+        expect(wrapper.find('button')).toHaveLength(3);
     });
 
     it('simulates click on each button', () => {
         const wrapper = shallow(getComponentWithDefaultProps());
         wrapper.find('button').map(b => b.simulate('click'));
-        expect(MockFn.mock.calls).toHaveLength(2);
+        expect(MockFn.mock.calls).toHaveLength(3);
     });
 });